FFFC-459 The overview for an account with no bills still has the pay this bill and recurring payment link disabled
Fixes this issue. Now follows the general model for active bills.  Pay this bill is enabled if payment is enabled for this user. AutoPay link is enabled if this user is not in DQ status, has payment enabled, and doesn't already have an autopay rule.

Refactored AccountOverview.uc to properly handle this situation in no bills and to removed "dead" code associated with invoice mode and tests for the balance calculation rules (which is very different in 1st Franklin).