SUPT1FFC-78 -- bug as updated... shows only one account when there are two (with no bill) and then marks payment against multiple accounts when paying that account.
Ok… fixed and tested on the same customer id and accounts on my local machine as those on UAT.  Problem was that I was using the wrong index variable when creating tm_account values for accounts with no bills.  These two accounts have no bills. The first one got registered in tm_account more than once.

The payment being applied to multiple accounts was an artifact of the fact that we had two records for the same account in tm_accounts.  It applied the payment once, but placed the same account number in the payment grouping twice!