A Sidebar defect ... associated with FFFC-771 -- eConsent Value Incorrectly Updated to False in Agent Application
Really should have been its own ticket.  Once a user agreed to eConsent again, the application used the viewStatus value and only accepted a status of enabled or showed access denied. We've added disabledEConsent as a valid value for showing since you'll never get to try to view the pages if eConsent is disabled (its blocked on login).